pyproject.toml
[project]
name = "ai-watermark-fighter"
version = "1.0.0"
description = "A Gradio application for protecting images from watermarks by adding and removing white canvas areas"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "AI Watermark Fighter Team" }
]
keywords = ["gradio", "image-processing", "watermark", "pillow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "gradio>=5.0.0,<6.0.0",
    "pillow>=10.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
ai-watermark-fighter = "cli:main"

[tool.hatch.build.targets.wheel]
packages = ["."]